Move TitleManager component to avoid circular import
